Tim development: Merged into master (#11)
* Tim's work (#9)

* updated SConstruct (vnmrbg) for clang build on OS X, buildovj for passing arguments (not used yet)

* Updated macos object files to be fat--i386 and x86_64; previously just x86_64. Compiled with clang -c -Os

* Added fidlib with fid1d.fid otherwise new installs cannot start up

* Image Magick's convert works (no links to missing dylibs)

* Don't use java link on Mac, use system java

* For OS X, pgsql is in /vnmr/pgsql, use it

* Use clang to compile. Option for 64 bit, but ALL components must be 64-bit, nt just vnmrbg

* Icons for OS X, remove broken convert

* moved ImageMagick to correct src/macos

* set date in src/common/vnmrrev

* added site_scons for building on OS X

* buildovj: detect # of cpus

* Removed duplicate OVJ build scripts buildovj, makeovj, and whatsin, which are now in ovjTools. Added .gitignores. (#10)

* Removed duplicate OVJ build scripts buildovj, makeovj, and whatsis
    * They are now in ovjTools

* Build artifacts were not being ignored by git status
    *Several .gitignore files added to suppress build artifacts from git status

+++ src/scripts/dbsetup.sh
@@ -1084,7 +1084,8 @@ echo "Using postgres path: $pgpath"
 if [ x$lflvr = "xdarwin" ]
 then
     # Remove launch file from postgres installation
-    oldfiles=`ls -1 /Library/LaunchDaemons/*postgresql-*` 
+    # TEB: should we really blow away another postgres?
+    oldfiles=`ls -1 /Library/LaunchDaemons/*postgresql-*`
     for file in $oldfiles; do 
         sudo rm -f $file 
     done 
@@ -1095,15 +1096,18 @@ then
         mkdir /vnmr/pgsql
         chown $vnmr_adm /vnmr/pgsql 
     fi
-    cd /vnmr/pgsql 
-    if [ -d bin ]; then 
-        if [ ! -d bin.sav ]; then
-            mv bin bin.sav 
-        else
-            rm -rf bin
+    if [ x$pgpath != x"/vnmr/pgsql/bin/" ]; then
+        cd /vnmr/pgsql
+        if [ -d bin ]; then 
+            if [ ! -d bin.sav ]; then
+                mv bin bin.sav 
+            else
+                rm -rf bin
+            fi
         fi
+        ln -s $pgpath bin
     fi
-    ln -s $pgpath bin
+    echo "login_user:.${login_user}. vnmr_adm:.${vnmr_adm}.\n"
 fi
 
 if [ x$osname = "xInterix" ]
